Separate p2p client code in other file
